email_service/requirements.txt
psycopg2
pydantic[email]
pika
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pydantic
pydantic[email]
python-dotenv
pybreaker